fai-rag-app/api-backend/pyproject.toml
[project]
name = "api-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "openai (>=1.61.1,<2.0.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "unstructured[docx,md,pdf] (>=0.16.23,<0.17.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
